requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via web3
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
bitarray==2.7.3
    # via eth-account
cached-property==1.5.2
    # via
    #   py-ecc
    #   py-evm
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
cytoolz==0.12.1
    # via eth-utils
eth-abi==4.0.0
    # via
    #   eth-account
    #   eth-tester
    #   web3
eth-account==0.8.0
    # via
    #   eth-tester
    #   web3
eth-bloom==2.0.0
    # via py-evm
eth-hash[pycryptodome,pysha3]==0.5.1
    # via
    #   eth-bloom
    #   eth-tester
    #   eth-utils
    #   trie
    #   web3
eth-keyfile==0.6.1
    # via eth-account
eth-keys==0.4.0
    # via
    #   eth-account
    #   eth-keyfile
    #   eth-tester
    #   py-evm
eth-rlp==0.3.0
    # via eth-account
eth-tester[py-evm]==v0.8.0-b.3
    # via web3
eth-typing==3.3.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   py-ecc
    #   py-evm
    #   web3
eth-utils==2.1.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   eth-tester
    #   py-ecc
    #   py-evm
    #   rlp
    #   trie
    #   web3
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
hexbytes==0.3.0
    # via
    #   eth-account
    #   eth-rlp
    #   trie
    #   web3
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jsonschema==4.17.3
    # via web3
loguru==0.7.0
    # via -r requirements.in
lru-dict==1.1.8
    # via
    #   py-evm
    #   web3
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.4
    # via
    #   py-ecc
    #   py-evm
packaging==23.1
    # via pytest
parsimonious==0.9.0
    # via eth-abi
pluggy==1.0.0
    # via pytest
protobuf==4.22.3
    # via web3
py-ecc==6.0.0
    # via py-evm
py-evm==0.6.1a2
    # via eth-tester
py-geth==3.12.0
    # via web3
py-solc-x==1.1.1
    # via -r requirements.in
pycryptodome==3.17
    # via
    #   eth-hash
    #   eth-keyfile
pyethash==0.1.27
    # via py-evm
pyrsistent==0.19.3
    # via jsonschema
pytest==7.3.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
regex==2023.3.23
    # via parsimonious
requests==2.28.2
    # via
    #   py-solc-x
    #   web3
rlp==3.0.0
    # via
    #   eth-account
    #   eth-rlp
    #   eth-tester
    #   py-evm
    #   trie
safe-pysha3==1.0.3
    # via eth-hash
semantic-version==2.10.0
    # via
    #   eth-tester
    #   py-geth
    #   py-solc-x
sortedcontainers==2.4.0
    # via trie
toolz==0.12.0
    # via cytoolz
trie==2.1.0
    # via py-evm
urllib3==1.26.15
    # via requests
web3[tester]==6.2.0
    # via -r requirements.in
websockets==11.0.1
    # via web3
yarl==1.8.2
    # via aiohttp
